Add files to support generic tooltips.
To use:
Add class 'tooltip' and a title attribute to any container. The title will then be used as the text of the tooltip.

Default tooltip position is above and centred on the container. Additional calsses can be added to control the tooltip position: ttleft, ttcentre, ttright, tttop, ttmiddle and ttbottom

git-svn-id: svn://suivarro/touchstone/trunk@528 8d9c46d6-3bce-41d5-b1ab-56de0fb72b2f