ROGO-622 - Notices generated by frequency discrimination analysis report
If a question does not have any answers for some reason (i.e. it was excluded by the late log, or the question was added after the users took the exam) then the report would generate notices.

It was also possible that the report might generate divide by zero errors.