ROGO-658: Calculation questions to not like bi-directional papers
If a user went to page with calculation question B that linked variables generated in calculation question A without visiting the page that question A is on first, then the linked variable would generate as 'ERROR'. This is then stored in the database. Question B will be unanswerable.

When the user goes to question A the variable will be generated, if they went to question B again the previously calculated variables would have been loaded from the database (i.e. 'ERROR') No attempt was made to get a working variable again. So question B was still unanswerable.

This change causes variable substitution to happen again if it is in an error state is found as a variable in a calculation question, which means question B should become answerable after visiting question A, even is question B was visited first.