ROGO-1499: Working behat test site
The behat test site will now run from a url and port combination defined in the Rogo configuration file using the values:

* $cfg_behat_website
* $cfg_behat_db_database
* $cfg_behat_db_user
* $cfg_behat_db_password

These values for the database and webserver should be different to the ones for the live site.

The behat site will use the same database user accounts to connect as the 'live' instance of Rogo, it is therefore important the behat site is not installed on a production server. The behat install also requires that a live site has been installed on the server in some form so that these database users already exist.

A rogo data directory for behat testing must be defined as $cfg_behat_data this directory must exist and the webserver must have permission to read and write to it.

All the predefined users will have a password that matches their username.