ROGO-1554: Class totals report displays users who are not in module
The visibility of users in the class totals report is set to assume that they should be displayed. There is then nothing in the code to set them to be hidden (although there is code to set them to be displayed.

This patch changes the default to assume the users should not be shown. this means that if their results are filtered out the user will not be displayed in as part of the report.