User data directory class system
In order to pave the way for configurable user data directories Rogo needs a way to access files indirectly. This commit introduces this by enabling a way to define directories and generate URLs for files in them, along with finding the files actual location and sending the file to the user.

In most cases to create a new user data directory will be as simple as creating a new class extending rogo_directory and implementing the two abstract methods.