The location of user information directories can now be configured.
To configure the location of the directories a $cfg_rogo_data setting should be defined inside the Rogo configuration file. All of the user data directories that use the rogo_directory::base_directory() method when defining their location will then move to use it.

The $cfg_rogo_data parameter will be ignored if the question interface is set to flash. As some of the interface files look for their data in paths that are relative to the Rogo root web directory.

The path in $cfg_rogo_data should be configured using the OS specific directory separator, on linux a / on windows it will be \, but because this is also the escape character it will need to be escaped as \\