ROGO-1573: Calendar not showing all summative papers in unfiltered view
It was possible for the extra calendar dates to overwrite the details of a Rogo paper. This could occur if there was a summative paper that has a property_id less than the number of extra calendar dates in Rogo added to the property_id of the paper with the latest start date of a summative system.

This change sets the id that the extra calendar dates use to start one after the maximum property_id of a summative paper.