ROGO-1499: Allow behat to define and run unit tests
Behat can hook into PHPUnit. We can now define tests to be run using PHPUnit rather than Mink by tagging them with @backend

The rogo_unit_test class lets the tests call PHPUnit assertions as though they are part of it. It also implements a data store to capture any class properties that are set and not defined. This allows us to pass data between steps in a way that is isolated between scenarios if we want.