Allow development systems to always run updates needed updates
Development systems may need to run update files even if if the configured version of Rogo matches the version the update is intended for.

UpdaterUtils::check_version() will now look for:

* $cfg_dev_system = true;

in the configuration file. If found the updates that are for a version number that matches the version number in the Rogo config file should also run.