ROGO-1489: Stop an SQL error when a users with no team search for papers
If a user who is not in any teams and not a SysAdmin searched for a paper in Rogo tehy would have had errors and no results as the SQL query generated for them was invalid.

This change ensures that such users are warned that they will not get any results until they are added to a team in Rogo.