ROGO-1744 Escape user input on search pages
Right now \ is likely to be recognised as the escape character in a test search box, to find a paper with the \ character in the title it would need to be manually escaped by the user. This change makes it so that the test input be the user is escaped in the script so they do not need to manually escape their search text.