ROGO-1787 User picture may not change
Before this patch if a student had a jpg user image uploaded and then a different profile png image was uploaded the jpg image would still display.

This patch fixes this case by deleting the existing user profile image for a user before uploading a new one. This ensures that they will have at most a single image in the profile directory.