When installing composer dependencies the lock file should be respected
Before this patch all of the composer dependencies would be updated to the latest version every time the init script for automatic tests was run. It will now read the lock file and get those versions unless the new --update parameter is used, or a new dependency is added.

This should make changes to the lock file happen in a more controlled way.