ROGO-1822 Labeling questions may not work correctly in the frequency and discrimination report
It looks as though in around 10 percent of labeling questions the stored correct answers for the question are stored with at least one floating point coordinate.

The user answers always seem to store the coordinates of an option as integers.

This report matches the user answer to the correct answers using the coordinates, so if there is a mismatch in the data for a question the number are wrong.

This patch ensures that within the report code, no matter how the answers are stored that they are converted to an integer that will match the value in the user answer.