Require version number match for behat tests
Before this change behat would run even if the version number in the config file was different to that of the version number of the code. This resulted in all the tests displaying the 'Rogo needs an upgrade' page rather than the page needed for testing.

Now behat will no longer attempt to run and will give a message that explains why.