ROGO-1363 Serialised objects with database connections causing errors
The error detail page in Rogo will produce errors if a serialised object that contains a database connection is part of the variables, as the mysqli objects will be invalid. This change causes the database variable of two objects that are know to cause this issue so far by setting the database connection property to null.

This change should remove the database object from any core classes when they are unserialised.

One of the logging methods was not using a built in function that removed certain objects from the trace.

These changes should stop most of these errors.