ROGO-1915 Paper reference buttons should display correctly when page resized
Reference buttons at the bottom of the page were positioned relative to the top of the page, this meant that when resized they disappeared off the bottom and were unusable. This change positions them relative to the bottom of the page so that when the window is resized they move and remain usable.