ROGO-1991 Notice about missing  uansunit index caused by Calculation questions
It happens if the question variables have been stored but no answer has been stored. I'm not sure how you can get into this state using the Rogo UI, but it occured lots during load testing.

This change ensures that when the index is not present no code tries to use it.