ROGO-2077 Students should not be able to start an exam after the end time
In summative exams this causes the student to only have late log entries that cannot be assigned to them.

This change continues to allow students to save an already started exam for upto 60 minutes after the end time, but will not allow them to start the exam after the end time.