ROGO-2153 Review comments lost in Rogo
Before this change if a reviewer went back to a paper after a one of their comments was responded to and they either changed page or pressed Finish.

This behaviour was caused because the save comment function deleted all the comments and then inserted new ones whenever a screen was changed.

After the change existing comments will be updated, with only new comments being inserted.