ROGO-2166 Fill in the blank questions with html in the blank options break
This change is designed to handle fill in the blank questions that have options that include html in the options. This causes a problem when the dropdown list method is enabled because the options in the list include the html in the submitted value. When submitted the html is stripped out causing the user's answer not to match any of the answers.

For example:
[blank]Option 1,<span>Option 2</span>,Option 3[/blank]

would result in a option similar to:
<option value="<span>Option 2</span>">...</option>

If a user selected the Option 2 the result save code would have changed the value to be simply: Option 2

This resulted in it not being selected as the default if the student went back to the exam page the question was on.

This change considers the following cases:

* Answers submitted before the filtering changes (i.e. stored as '<span>Option 2</span>')
* Answers submitted after the filtering change (i.e. stored as 'Option 2')