ROGO-2168 Unanswered parts in fill in the blank questions can error
If a Blank Textbox fill in the blank question had multiple parts the unanswered parts parameters evaluated to null, this meant that the part was ignored and not saved. Which in turn could mean that the saved response array would contain less values than expected by the question.

This would result in answers possibly appearing to move.

This change ensures that when saving every part of the question gets a stored response.