ROGO-2256 Answered summative can questions unlock if used on another paper
Before this change the method only checked if the question had been answered on a single paper.

After this change the method will check if the question has been answered on any summative paper.

This change is required to stop questions being automatically unlocked if they are linked to both a used summative exam and a new one that has not yet run when a teacher visits the new paper.