ROGO-2264 Modules with no enroled users results in invalid SQL
When the assessment export is run on a paper for a specific module, if there are no enroled students on the year that the paper is for then an empty in statement would be generated and used in a later query.

An empty IN statement is invalid SQL. Now when this happens the SQL will look for userID with a value of NULL, since this should never be the case no results will be found, which is what we would expect.

We cannot remove the SQL entirely as that would potentially cause results for students that are not on the module to be displayed.