Merged in hotfix/ROGO-2553 (pull request #1014)
ROGO-2553 Underscores in usernames not handled well by create user form

* ROGO-2553 Display warning if username is invalid

Before this patch if a username contained an underscore the form the user would be given a blank screen on submitting the form.

* ROGO-2553 Form should not forget data

If a username was not unique or contained an underscore the create user page would require the user to reselect the Status and Type/Course for a user. They will now display the value the user had selected when they submitted the form.

Approved-by: Joseph Baxter
Approved-by: yijun xue