Merged in improvement/ldap (pull request #1059)
ROGO-2549 Make LDAP great again!

* ROGO-2549 Improve documentation and cleanup white space

* ROGO-2549 Allow more complicated username queries

This change will allow LDAP to search for a user where a query involving multiple paramerters is needed for example:

(&(type=Person)(username=bob))

The ldap_user_prefix would be: (&(type=Person)(username=

The ldap_user_postfix would be: ))

This would be useful where just using username=bob is slow,returned multiple entries, or you wished to restrict access to users who are in a group.

* ROGO-2549 We might try to create duplicate users if an authentication plugin doe not use the core users table

* ROGO-2549 Empty filters do not work

Trying to run an empty search filter results in a disconnetion for the LDAP server.

* ROGO-2549 Refactor to make code easier to read

It also improves the documentation

Approved-by: Joseph Baxter
Approved-by: yijun xue