Merged in hotfix/ROGO-2561 (pull request #1066)
ROGO-2504 Keywords should not be lost on intermediate save screen

* ROGO-2504 Keywords should not be lost on intermediate save screen

Before this change if you edit the keywords in a question with an intermediate screen, for example Hotspot, when it is locked and make a change to the Keywords on the first editing page they disappear when you save and move to the intermediate page.

After this change the keywords are saved when the intermediate page is loaded so they are not lost.

* ROGO-1613 Filter objective mappings

* ROGO-2504 Some functionality will not work as expected when page not English

$mode is set to be either the loaclised add or edit string.

Approved-by: Joseph Baxter
Approved-by: yijun xue