ROGO-2541 Make it clear when a database connection issue affects installation
Before this change if Rogo cannot connect to the database because it's details or the user does not have permission the only message was that MySQL did not meet the minimum requirements. This was obviously confusing when they had a supported version running.

After the change they will be informed that a connection could not be established along with the reason from MySQL.