ROGO-2048 Display all labs in system
If a lab has no clients assigned to it, then it was not displayed before this change.

The Lab details page needed modifying to handle this situation.

It should not be possible to create new labs with no clients, but it was obviously possible in the past, we should make them visible so that they can be given clients or deleted via the Rogo UI.