Merged in improvement/grunt-js-warnings (pull request #1168)
ROGO-2313 Fix all no-undef warnings

* ROGO-2313 Define more globals

We do not want our linting to warn us about these

* ROGO-2313 Do not lint 3rd party code

* ROGO-2313 Required module not defined

* ROGO-2313 Variables not referenced correctly

These variables were being used as globals, but they are properties of the ref prototype passed to the method.

* ROGO-2313 Remove code that does nothing

I do not think the test mode is used, but this global is not used anywhere else.

* ROGO-2313 Stop using global variables

These variables should not be global!

* ROGO-2313 Remove code that will never execute

showtest is hardcoded as false so the entire block will never run. Removing it means there is no need to pass ref to the method.

* ROGO-2313 Scope variables correctly

* ROGO-2313 Do not minify code if it adds unauthorised globals

Globals are bad. Their presense generally mean (espcially now that we have requirejs) that the code is poorly written and may be buggy.

Approved-by: yijun xue
Approved-by: Joseph Baxter