ROGO-642 Student cohort of paper list
This patch allows teachers to see a list of students who can take a paper.

A link to it will be displayed in the Paper tasks menu if the paper has a year restriction set. There are two reasons that it is not allowed on papers where the year is not restricted:

1) I fear it could perform poorly and produce a massive list
2) The method in paper properties does not currently return any results if there is no year set

The list will filter students to people who are on the modules and meet any meta data restrictions on the paper.