ROGO-2765 Redirect to install page correctly in virtual directories
When Rogo is installed in a virtual directory the DOCUMENT_ROOT is unlikely to match any part of the root of a file in Rogo.

This change allows us to calculate the relative path from the file the user called to the root directory of Rogo even in virtual directories. This will allow us to always redirect correctly to the installation page.