ROGO-2776 Dump on failed behat steps
A user may configure a faildump directory in the behat.xml config file. If they do when a step fails a screenshot and html dump will be generated for the state of the browser.

The files will be placed inside a directory that uses the timestamp of the test start (so should usually be uniqueto a run)

The screenshot and html dump will be names for the feature file and line number