Merged in hotfix/ROGO-2815 (pull request #1365)
ROGO-2815 Highlight hovered options in MCQ and MRQ questions

* ROGO-2815 Highlight hovered options in MCQ and MRQ questions

This change is to fix issues with MCQ and MRQ questions when an option was on multiple lines.

Before this change it could be difficult to distinguish which check box or radio button was linked to which text, after it when hovering over an option ot text will highlight all of that option to make it easier to tell.

* ROGO-2815 Add highlights to the review screen

Approved-by: Joseph Baxter