ROGO-2662 Allow timer resets on users who have no results
This change ensures that a user who has started an exam, but not had any results entered onto it can have their timer reset.

It is to address edge cases such as a user doing a progress test having technical issues immediatly after pressing start (when no results would have been saved for them), before this change they would only have shown as absent and it was not possible to reset their time via the user interface.

After this change they will still only appear if absent candidates are selected but the option to reset the time on their exams will be present.

Additionally it will grey out the reset time link on truely absent candidates.