Merged in hotfix/ROGO-2936 (pull request #1416)
ROGO-2936 Fix links on search page

* ROGO-2936 Fix links on search page

Before this change the Copy to paper x and Link to paper X menu options on question search gave an access denied error.

This happened because a there is a qType value on the page that is not an empty string, but no module value so undefined was being passed, however since there is no module called undefined the users were being denied access to the link creation page.

Approved-by: Yijun Xue
Approved-by: Joseph Baxter