Merged in hotfix/ROGO-3052 (pull request #1464)
ROGO-3052 Stale files causing issues in exams after updates

* ROGO-3052 Stale files causing issues in exams after updates

This change will mean that javascript and css files get revalidated by user's browsers every 15 minutes, which should mean that exams would be highly unlikely to be affected by out of date files.

It will however mean that there will be extra requests sent to the server by students during exam. Enabling ETags should mean that minimal content is ever returned to the users as if there is no change to the tag the browser will not be sent a new copy of the file.

This change will only affect Rogo if it is being run on Apache server that is configured to process this file, all other servers will need to modify their own configuration for file serving.

Approved-by: Yijun Xue
Approved-by: Joseph Baxter