ROGO-3192 Tests should not fail on last day of the year
The tests in this file did not create the intended data on the last day of the year as the default start and end date were for the next day. This meant that on 31st December all papers were created in the next calendar year.

This change ensures that the papers are always created so that they are in the correct year by changing the ones that had the default values to be created for the current day.