ROGO-3199 Exam navigation should only be fully active after JS loads
The non-default navigation buttons for exams will now be disabled until JavaScript for the page has loaded enough to correctly handle submitting to the correct location.

We are doing this because students have been having issues where the pervious button has been moving them to the next screen and/or finishing the exam.