ROGO-3219 End session early when user downloads a file
When the session is open PHP will not process a second PHP file for the user, this means that if an exam script has many images before this change we could only download one at a time.

Now as soon as the user has been verified we end the session, which means another image download can start this should make Rogo pages load more quickly.