ROGO-3234 Stop deprecation messages caused by module checks
The checklist property is put through mb_substr to find out if it contains certain values, in PHP 8 it is deprecated to pass null to it rather than a string. Null is stored in the database when it has no value so we need to ensure that when getting the value it is converted to an empty string.