ROGO-3234 Stop change in behaviour
It looks like we expect media records, even when their is no media for an option. This change modifies a previous commit so that unit tests pass retaining the old behaviour while fixing a null into pathinfo is deprecated message.

Before merging this commit will be rebased into ROGO-3234 Protect against trying to set media that does not exist