EXAMSYS-3276 Should not see the previous attempts without completion
Before this change if a user had attempts that were not completed they would be presented a list of attempts that they could not access.

This change ensures that the user will only see the links to previous attempts if they will be able to access them.

The change makes this check use logic similar for displaying the previous papers link to that used on the paper finsih page