EXAMSYS-3134 Stop export error when users have not answered hotspots
If the rpt_letters_for_hotspots setting is enabled and a user did not answer the hotspot question by not going to the screen it is on then the assignment export would fail with an unhandled exception as a null was passed into the markWithLetters method.

It looks as though the first if statement here was simply being discarded with it's last part and the first statement of the second one being duplicated.

After this change markWithLetters should not get called if the user has not answered the question.