EXAMSYS-3385 Staff accounts should be able to edit summative examss
When summative management is enabled in ExamSys staff level accounts cannot edit the start and end times of papers and the form elements for them are disabled.

This means that values for the fields are not passed for them during a form submission, however the script was still requiring that the values were set.

This resulted in output being sent to Javascript that it was not expecting, which caused an error.