EXAMSYS-3384 There should be no errors for unmarked questions
If a question is unmarked the value of $tmp_user_mark will be null. This will happen in the case of short answer text box questions, and could in theory happen in the case of calculation questions.

This change handles cases where it is null to avoid errors being generated.