EXAMSYS-3377 Ensure that the random mark is always a number
The code in ExamSys assumes that the random mark is always a number, however there are occasions where it is stored as null in the database.

This change is designed to stop null from being stored for it in the database from now on, and to handle the existing records that use null.