EXAMSYS-3387 Handle users with no existing notes
When a user has no notes StudentNotes::get_note() returns false, this causes issues when generating the form further down as it tries to use false as an array.

This change gives some sensible defaults in this case.

I did consider changing the return value of StudentNotes::get_note() to be an array even when there is no note for the user on the paper, however other parts of ExamSys use the false value, which meant that this change has the minimum scope for going wrong.