EXAMSYS-3375 Stop errors when importing via QTi
Before this change if the imported question contained an image tag in it's presentation text, for example:

    <img src="myimage.png" alt="A nice picture">

errors would be generated.

This change causes the tag to be parsed in a slightly better way that will allow it to work for the only case it is currently used to find the src attribute of image tags.

It will probably be unusable for many other use cases.