EXAMSYS-3452 Change query to handle old date time format for table review_metadata
The table metadata_review has columns that were not upgraded as part of the change to
UNIX timecode. Line 680 contained code that assumed that change over was complete.
This fix removes the UTC assumption, and updates the script to convert the started
date to a format that is used for ordering, and another format that is used to display
the date in the interface, referred to as display_started.