EXAMSYS-3467 Make broken labels distractors
If a question has a label with broken co-ordinates we do not want it to be impossible to display so set it up as a distractor.

This will mean that it is not treated as a correct answer during marking, which is probably what is wanted as the users will not have been able to see it during editing before this change.