EXAMSYS-3429 Ensure that plain text answers decode html entities before being displayed to student
Before this change if an answer was saved to a plain text or mathjax textbox question and then the user went back to the same page in the exam any characters that should be encoded as html entities when displaying in a web page showed the entity code.

For example if & was used it would be displayed as &amp;

This happens because we store the answers using encoded html entities so that we do not need to encode them when outputting into pages such as marking and feedback screens.