EXAMSYS-3462 Fix issues with theme, scenario and notes being empty
If the theme, scenario or notes of a question is stored internally in the database as a null, rather than an empty string there are many pages in ExamSys that would have had an error.

Most of the code in the system assumes that they are an empty string when they have not been set.

I think now they can mainly have null values in this field if they were created by either RAF or QTi import.