EXAMSYS-3476 Likert questions with the N/A option enabled should work
Before this change when N/A was enabled for likert questions they would fail on surveys.

I have also modified the likert question output somewhat to make it selectable in automatic tests.

In general in behat likert questions can have their options created in a human-readable way, we can now also answer them using the answer that is displayed to the user.

The OSCE test paper now uses an OSCE scale.