EXAMSYS-3465 Improve robustness of calculation questions
Before this change on all engines if there was a variable that did maths on another variable which was linked to another question there would be an unhandled exception when the linked question did not have a stored value for it, for example if the question was on the same page or a user had not answered it.

The phpEval engine would also break ExamSys if there was a parsing error in either a answer or variable formula.

After this change calculation questions should not stop ExamSys from continuing to generate a page if there is an error in them.