EXAMSYS-3489 We should not try to create a new variable
Before this change the code would have tried to create a new variable witht he name of the output of the cosign_debug method.

The cosign_debug method returns nothing, so I can only conclude that this code is incorrect.

No other call to cosign_debug in this method tried to convert its output into a variable name.