EXAMSYS-3489 Fix formatting of constructors
When class properties that had PHP Docs were promoted it was done in a way that did not conform to the ExamSys coding standard.

There were also some cases where details from the original property were not transferred into the constructor description, in these cases I have written hopefully a better version of the docs.