EXAMSYS-3489 Improved display error installation method
All uses of the method pass an array (anything else would not have been displayed)

We no longer need to test if an array is being passed, however for web rendering we do need test if it contains any values to retain the same behaviour.

The PHP docs for the method should now be better.