EXAMSYS-3510 Remove unused code
This block of code used to be used to create part of the url for the paper properties page.

It is no longer used for that, however it does overwrite the value of the `$module` string. This is the second place in the file where the variable could be set.

This one could be set to null which would cause fatal errors on summative papers when:

* an internal or external review is configured
* it is attached to multiple modules
* the url for the page does not reference a specific module

Since this code is no longer used it appears to be safe to remove it.